% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-coerce.R
\name{as}
\alias{FLStock2biodyn}
\alias{as}
\title{as("biodyn", "FLStock")}
\usage{
FLStock2biodyn(from, control = biodyn()@control, mult = FLPar(rep(c(0.5, 1,
  1/0.5), each = 4), dimnames = dimnames(biodyn()@control[, -1])))
}
\description{
Coerces an \code{FLStock} into a \code{biodyn} object.
}

