% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspic-class.R
\docType{class}
\name{aspic-class}
\alias{aspic-class}
\alias{catch,aspic-method}
\alias{catch<-,aspic,FLQuant-method}
\title{ASPIC Biomass Dynamic Model Class}
\value{
aspics object
}
\description{
A class that represents the ASPIC biomass dynamic stock assessment model.
}
\details{
The Class is intended to be used as part of an MSE and includes methods for diagnostics, calculating  reference points and other quantities used when providing management advice.
}
\section{Slots}{

\describe{
\item{\code{name}}{{\code{character} with the name of the stock}}

\item{\code{conditioning}}{{ \code{factor}}}

\item{\code{options}}{{ \code{numeric}}}

\item{\code{index}}{{ \code{data.frame}}}

\item{\code{stopmess}}{{ \code{character}}}

\item{\code{rnd}}{{ \code{numeric}}}

\item{\code{model}}{{ \code{factor}}}

\item{\code{catch}}{{ \code{FLQuant}}}

\item{\code{stock}}{{ \code{FLQuant}}}

\item{\code{diags}}{{ \code{data.frame}}}

\item{\code{params}}{{ \code{FLPar}}}

\item{\code{control}}{{ \code{FLPar}}}

\item{\code{priors}}{{ \code{array}}}

\item{\code{vcov}}{{ \code{FLPar}}}

\item{\code{hessian}}{{ \code{FLPar}}}

\item{\code{objFn}}{{ \code{FLPar}}}

\item{\code{mng}}{{ \code{FLPars}}}

\item{\code{name}}{{ \code{character}}}

\item{\code{desc}}{{ \code{character}}}

\item{\code{range}}{{ \code{numeric}}}
}}
\examples{

\dontrun{aspic()}

}

