% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspic-fit.R
\docType{methods}
\name{boot}
\alias{boot}
\alias{fit,aspic,missing-method}
\title{fit}
\arguments{
\item{object;}{an \code{aspic} object or a character string giving an aspic "inp" file}

\item{dir;}{an optional \code{dir} where aspic text files used for fitting can be found}
}
\value{
An aspic object with fitted values and parameter estimates

An aspic object with fitted values and parameter estimates
}
\description{
Fits the aspic model to catch and catch per unit effort data

Bootstraps the ASPIC biomass dynamic model.

Bootstraps the aspic model
}
\examples{
\dontrun{
    data(asp)
    asp=boot(asp)}
}
\seealso{
\code{\link{aspic},\link{biodyn},\link{boot},\link{jk}}

\code{\link{biodyn},\link{boot},\link{jk}}
}
