% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspic-xval.R
\name{xval,biodyn,missing-method}
\alias{xval,biodyn,missing-method}
\title{xval}
\usage{
\S4method{xval}{biodyn,missing}(
  object,
  minyear = range(object)["maxyear"] - 4,
  maxyear = range(object)["maxyear"]
)
}
\arguments{
\item{object}{an object of class \code{biodyn}}

\item{minyear}{last year to fit data to}

\item{maxyear}{last year to project to, by default is the last year in the catch}

\item{index}{an \code{FLQuant} or \code{FLQuants} with index of relative stock abundance}
}
\description{
Performs a cross-validation uisng a hindcast
}
\details{
xval

Returns a data.frame with index, year, obs and hat
}
\examples{
\dontrun{
 data(bd)  
}
}
