% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{aspic}
\alias{aspic}
\alias{aspic,ANY,ANY-method}
\alias{aspic,ANY-method}
\alias{aspic,character-method}
\alias{aspic,data.frame-method}
\alias{aspic,missing-method}
\alias{aspic-method}
\title{aspic constructor}
\usage{
aspic(object, value, ...)
}
\arguments{
\item{object}{can be \code{character} or \code{biodyn} objects}

\item{...}{named parameter being passed to slots}
}
\value{
aspic object
}
\description{
Create an object of the \strong{aspic} class representing a biomass dynamic stock assessment model.
}
\examples{
\dontrun{
asp=aspic()
}
}

