% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{aspics}
\alias{aspics}
\alias{aspics,list-method}
\alias{aspics,missing-method}
\alias{aspics-method}
\title{aspics}
\usage{
aspics(object, ...)
}
\arguments{
\item{object}{can be \code{aspic} object or a \code{list} of \code{aspic} objects}

\item{...}{additional \code{aspic} objects}
}
\value{
\code{aspics} object
}
\description{
Create a list of aspic objects
}
\examples{
\dontrun{
aspics(aspic())
}
}

