% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/aspics.R
\docType{class}
\name{aspics}
\alias{aspics}
\alias{aspics-method}
\alias{aspics,missing-method}
\alias{aspics,aspic-method}
\alias{aspics,list-method}
\alias{aspics-class}
\alias{aspics-methods}
\alias{aspics,ANY-method}
\alias{aspics,character-method}
\title{aspics}
\usage{
aspics(object, ...)

\S4method{aspics}{list}(object, ...)
}
\arguments{
\item{object}{can be \code{aspic} object or a \code{list} of \code{aspic} objects}

\item{...}{additional \code{aspic} objects}
}
\value{
\code{aspics} object
}
\description{
Create a list of aspic objects

features that give a little more structure to list objects. First the
elements of \code{aspics} must all be of the same class \code{biodyn}. 
Second it implements lock mechanism that, when turned on, does not allow 
the user to increase
or decrease the object length.
}
\section{Slots}{
 \describe{
  \item{.Data}{The data. \code{list}.}
  \item{names}{Names of the list elements. \code{character}.}
  \item{desc}{Description of the object. \code{character}.}
  \item{lock}{Lock mechanism, if turned on the length of the list can not be
    modified by adding or removing elements. \code{logical}.} }
}

\examples{
\dontrun{
aspics(aspic())
}
\dontrun{
asp <- aspics("1"=aspic(),"2"=aspic())}

}
\seealso{
\link[base]{[}, \link[base]{[<-}, \link[base]{[[<-},
\link[base]{$<-}, \link[methods]{coerce}, \link[base]{lapply},
\link[stats]{window}, \link[base]{list}
}
\author{
The FLR Team
}
\keyword{classes}
