% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-accessors.R
\docType{methods}
\name{control<-}
\alias{control<-}
\title{control<-}
\usage{
control(object) <- value
}
\value{
\code{biodyn} with new control slot
}
\description{
sets in \code{biodyn} initial guess and lower and upper bounds
}
\examples{
 
\dontrun{control(biodyn())}
}

