% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-class.R
\docType{class}
\name{biodyn-class}
\alias{biodyn,FLQuant,FLPar-method}
\alias{biodyn,FLQuant,missing-method}
\alias{biodyn,FLStock,FLPar-method}
\alias{biodyn,FLStock,missing-method}
\alias{biodyn,character,FLPar-method}
\alias{biodyn,character,missing-method}
\alias{biodyn,factor,FLPar-method}
\alias{biodyn,factor,missing-method}
\alias{biodyn,missing,FLPar-method}
\alias{biodyn,missing,missing-method}
\alias{biodyn-class}
\title{biodyn class}
\description{
A class that implement a biomass dynamic stock assessment model.
}
\details{
The Class is intended to be used as part of an MSE and includes methods for diagnostics, calculating  reference points and other quantities used when providing management advice.
}
\section{Slots}{

\describe{
\item{\code{name}}{{A \code{character} with the name of the stock}}

\item{\code{desc}}{{A \code{character} providing a fuller description of the object}}

\item{\code{range}}{{A \code{numeric} vector containing the quant and year ranges}}

\item{\code{model}}{{A \code{factor} giving name of production function, for now this is only `pellat`}}

\item{\code{obj}}{{ \code{factor} that determines the objective function type -LL or LAV}}

\item{\code{catch}}{{An \code{FLQuant} with total catch by year}}

\item{\code{stock}}{{An \code{FLQuant} which will hold the estimated stock by year}}

\item{\code{control}}{{An \code{FLPar} which sets initial guess (val) and bounds (min and max) for each parameter. The phase allows a parameter to be fixed if less <0 and for paramters to be estimated sequentially}}

\item{\code{hcr}}{{A \code{data.frame} with harvest control rule options}}

\item{\code{priors}}{{An \code{array} which sets penalties for parameters}}

\item{\code{params}}{{An \code{FLPar} with parameter estmates}}

\item{\code{vcov}}{{An \code{FLPar} with the covariance matrix of the parameter estimates}}

\item{\code{hessian}}{{An \code{FLPar} with the hessian of the estimated parameters}}

\item{\code{ref}}{{A \code{numeric} with parameters for estimating mng quantities}}

\item{\code{mng}}{{\code{FLPar} with derived quatities of management interest}}

\item{\code{mngVcov}}{{An \code{FLPar} with the variance matrix of management quanties}}

\item{\code{diags}}{{A \code{data.frame} with residuals and covariates from fit of CPUE to stock }}

\item{\code{objFn}}{{\code{FLPar} with objective function}}

\item{\code{ll}}{{\code{FLPar} with negative log likelihood by data component}}

\item{\code{profile}}{{\code{data.frame} not yet implemented}}

\item{\code{hst}}{{\code{data.frame} not yet implemented} 

All slots in the class have accessor and replacement methods that provide validation and protection of their data.}
}}
\examples{
\dontrun{biodyn()}
}

