% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-control.R
\docType{methods}
\name{control<-,biodyn,FLPar-method}
\alias{control<-,biodyn,FLPar-method}
\title{control<-}
\usage{
\S4method{control}{biodyn,FLPar}(object) <- value
}
\value{
\code{biodyn} with new control slot
}
\description{
sets in \code{biodyn} initial guess and lower and upper bounds
}
\examples{
 
\dontrun{control(biodyn())}
}
