% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{biodyn}
\alias{biodyn}
\title{biodyn constructor}
\usage{
biodyn(object, params, ...)
}
\arguments{
\item{object}{can be \code{character} object or a \code{FLPar}, \code{FLStock}, \code{aspic} objects}

\item{...}{named parameter being passed to slots}
}
\value{
biodyn object
}
\description{
Creates an object of the biodyn class representing a biomass dynamic stock assessment model.
}
\examples{
\dontrun{
bd=biodyn(params=FLPar(r=0.6,k=50000,p=1,b0=1))
}
}

