% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyns.R
\docType{methods}
\name{biodyns,biodyn-method}
\alias{biodyns,biodyn-method}
\title{biodyns creates a list of biodyn objects}
\usage{
\S4method{biodyns}{biodyn}(object)
}
\description{
biodyns creates a list of biodyn objects
}
\section{Slots}{

\describe{
\item{\code{object}}{}
}}

