% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{data}
\name{biodyn}
\alias{biodyn}
\alias{biodyns,biodyn-method}
\alias{biodyns,list-method}
\alias{biodyns,missing-method}
\alias{biodyns-method}
\title{biodyns}
\format{An object of class \code{NULL} of length 0.}
\arguments{
\item{object}{can be \code{biodyn} object or a \code{biodyn} of \code{biodyn} objects}

\item{...}{additional \code{biodyn} objects}
}
\value{
\code{biodyns} object
}
\description{
Create a list of biodyn objects
}
\examples{
\dontrun{
biodys(biodyns())
}
}
\keyword{datasets}

