% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyns.R
\docType{class}
\name{biodyns}
\alias{biodyns}
\alias{biodyns,ANY-method}
\alias{biodyns,list-method}
\alias{biodyns,missing-method}
\alias{biodyns-class}
\alias{biodyns-methods}
\title{biodyns constructor}
\usage{
\S4method{biodyns}{list}(object, ...)
}
\description{
The \code{biodyns} is a class that extends \code{list} but implements a set of
features that give a little more structure to list objects. First the
elements of \code{biodyns} must all be of the same class \code{biodyn}. 
Second it implements lock mechanism that, when turned on, does not allow 
the user to increase
or decrease the object length.
}
\section{Slots}{
 \describe{
  \item{.Data}{The data. \code{list}.}
  \item{names}{Names of the list elements. \code{character}.}
  \item{desc}{Description of the object. \code{character}.}
  \item{lock}{Lock mechanism, if turned on the length of the list can not be
    modified by adding or removing elements. \code{logical}.} }
}
\examples{
\dontrun{
bds <- biodyns("1"=sim(),"2"=biodyn())}

}
\author{
The FLR Team
}
\seealso{
\link[base]{[}, \link[base]{[<-}, \link[base]{[[<-},
\link[base]{$<-}, \link[methods]{coerce}, \link[base]{lapply},
\link[stats]{window}, \link[base]{list}
}
\keyword{classes}

