% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biodyn-msy.R
\name{bmsy}
\alias{bmsy}
\alias{bmsy,biodyn,missing-method}
\alias{bmsy,biodyn,numeric-method}
\alias{bmsy,character,FLPar-method}
\alias{bmsy,factor,FLPar-method}
\title{bmsy}
\usage{
bmsy(object, params, ...)
}
\arguments{
\item{object}{an object of class \code{biodyn} or a string or factor that species the model}

\item{params}{an \code{FLPar} object with model parameters}

\item{...}{any other parameters}
}
\value{
an \code{FLPar} object with value(s) of MSY, B_MSY and B_MSY
}
\description{
Calculates BMSY given the model parameters, can be done for a biodyn class, or by specifying the model and parameters
}
\examples{
\dontrun{
refpts('logistic',FLPar(msy=100,k=500))
}
}
\seealso{
\code{\link{msy}}, \code{\link{fmsy}} and  \code{\link{bmsy}}
}

