% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspic-fit.R, R/biodyn-boot.R
\docType{methods}
\name{boot}
\alias{boot}
\alias{boot,biodyn-method}
\alias{boot-method}
\title{boot, Bootstraps the ASPIC biomass dynamic model.}
\usage{
boot(object, ...)

boot(object, ...)
}
\arguments{
\item{object;}{an \code{aspic} object or}

\item{object;}{a character string giving an aspic "inp" file}

\item{dir;}{an optional \code{dir} where aspic text files used for fitting can be found}

\item{object;}{a \code{biodyn} object}
}
\value{
An aspic object with fitted values and parameter estimates
}
\description{
Bootstraps the aspic model

Bootstraps the aspic model
}
\examples{
\dontrun{
    data(asp)
    asp=boot(asp)}
\dontrun{
    data(asp)
    asp=boot(asp)}
}
\seealso{
\code{\link{biodyn},\link{boot},\link{jk}}

\code{\link{biodyn},\link{boot},\link{jk}}
}

