% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchEntropy.R
\name{branchEntropy}
\alias{branchEntropy}
\title{The branch entropy of Chinese words with two characters.}
\usage{
branchEntropy(two_chars)
}
\arguments{
\item{two_chars}{Chinese word with two characters.}
}
\value{
branch entropy of target word.
}
\description{
This function ccalculate both inner and outer entropy of candidate words.
chinese character frequency Statistics table.
}
\examples{
\donttest{
further examples for users (not used for checks)
branchEntropy("test character")
}
}
