% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charEntropy.R
\name{charEntropy}
\alias{charEntropy}
\alias{innerH}
\title{The branch entropy of Chinese words with two characters.}
\usage{
innerH(two_char, segWords)
}
\arguments{
\item{candidateWord}{A single Chinese character}

\item{topN}{The number of raws of output dataframe}
}
\value{
A dataframe containing the next word and its occurrence and frequency
}
\description{
This is a function that lists the next word appearing frequency
by a given chinese character based on real world
chinese character frequency Statistics table.
}
\examples{

}
