% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/biodyn-accessors.R
\name{setControl<-}
\alias{control}
\alias{control,biodyn,missing-method}
\alias{control-method}
\alias{setControl<-}
\alias{setParams<-}
\alias{setParams<-,aspic,data.frame-method}
\alias{setParams<-,biodyn,FLBRP-method}
\alias{setParams<-,biodyn,FLPar-method}
\alias{setParams<-,biodyn,FLQuant-method}
\alias{setParams<-,biodyn,FLQuants-method}
\alias{setParams<-,biodyn,data.frame-method}
\title{control}
\usage{
setControl(object, ...) <- value

setParams(object, ...) <- value

control(object, ...)
}
\arguments{
\item{object}{either \emph{biodyn} or \emph{aspic} class}

\item{...}{any other parameter}

\item{value}{}

\item{object}{either \emph{biodyn} or \emph{aspic} class}

\item{value}{}

\item{...}{any other parameter}
}
\value{
FLPar
}
\description{
Slot for fitting options, i.e. for starting guesses, to fix parameters or set bounds

Slot for estimates

sets initial guess and lower and upper bounds
}
\examples{
\dontrun{control(biodyn())}
}

