% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-control.R
\name{controlFn}
\alias{controlFn}
\title{controlFn}
\usage{
controlFn(object)
}
\arguments{
\item{r}{a \code{numeric} value with best guess}

\item{k}{a \code{numeric} value with best guess}

\item{p}{a \code{numeric} value with best guess, default=1}

\item{b0}{a \code{numeric} value with best guess default=1}

\item{phaseR}{a \code{numeric} value for phase, default=1}

\item{phaseK}{a \code{numeric}  value for phase, default=1,}

\item{phaseP}{a \code{numeric}  value for phase, default=-1,}

\item{phaseB0}{a \code{numeric}  value for phase, default=-1,}

\item{min}{a \code{numeric} a multipler for the best guess}

\item{max}{\code{numeric} a multipler for the best guess}
}
\value{
a \code{control} object
}
\description{
A utility function to help set up the \code{control} slot in \code{biodyn}
}
\examples{
\dontrun{
sim()
   }
}
\seealso{
\code{\link{biodyn}}  \code{\link{control}}
}

