% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drugnames}
\alias{drugnames}
\title{drugnames}
\format{A data frame with 1 variables:
\describe{
\item{\code{V1}}{Target string.}
}}
\usage{
drugnames
}
\description{
This dataframe contains examples of user-imported data.
}
\keyword{datasets}
