% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drugnames_test}
\alias{drugnames_test}
\title{This dataframe contains examples of user-imported data(used for test).}
\format{A data frame with 1 variables:
\describe{
\item{\code{V1}}{Target string.}
}}
\usage{
drugnames_test
}
\description{
This dataframe contains examples of user-imported data(used for test).
}
\keyword{datasets}
