% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-feasible.R, R/biodyn-tmb.R
\docType{methods}
\name{feasible}
\alias{feasible}
\alias{feasible,biodyn-method}
\alias{feasible-method}
\alias{nll}
\title{feasible}
\usage{
feasible(object, catch, ...)

feasible(object, catch, ...)

nll(object, index, ...)
}
\arguments{
\item{object}{an \code{FLPar}, \code{FLBRP} or \code{data.frame} object with parameters for the production function, r, k, p and b0.}

\item{catch}{an \code{FLQuant} with a time series of catch, iters must be equal to 1}

\item{object;}{a \code{biodyn} object}

\item{object;}{a \code{biodyn} object}
}
\value{
a \code{FLPar} a subset of params with parameter values that can explain the catch
}
\description{
Checks that the parameters can be taken given the catch time series

Checks that parameters are feasible by ensuring catch can be taken given parameters.

Checks that parameters are feasible by ensuring catch can be taken given parameters.
}
\examples{
\dontrun{
    dat(asap)
    asp=boot(asp)}
\dontrun{
    data(asp)
    asp=boot(asp)}
}
\seealso{
\code{\link{biodyn}}

\code{\link{biodyn}}
}

