% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biodyn-fit.R
\name{diagsFn}
\alias{diagsFn}
\title{fit}
\usage{
diagsFn(res)
}
\arguments{
\item{object}{an object of class \code{biodyn}}

\item{index}{an \code{FLQuant}, \code{FLQuants} or  \code{data.frame} object with CPUE indices}

\item{...}{other arguments}
}
\description{
Estimates parameters \code{biodyn} class by fitting catch and CPUE indices
}
\examples{
\dontrun{
   #simulate an object with known properties
   bd=sim()
   bd=window(bd,end=49)

   #simulate a proxy for stock abundance
   cpue=(stock(bd)[,-dims(bd)$year]+stock(bd)[,-1])/2
   cpue=rlnorm(1,log(cpue),.2)

   #set parameters
   setParams(bd) =cpue
   setControl(bd)=params(bd)
   control(bd)[3:4,"phase"]=-1

   #fit
   bd=fit(bd,cpue)
}
}

