% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/aspic-fit.R
\docType{methods}
\name{fit}
\alias{fit}
\alias{fit,aspic,missing-method}
\alias{fit,aspic-method}
\alias{fit,aspics-method}
\alias{fit,biodyn,FLQuant-method}
\alias{fit,biodyn,FLQuants-method}
\title{fit}
\usage{
fit(object, index, ...)

\S4method{fit}{aspic,missing}(object, dir = tempdir(), package = "mpb",
  exeNm = "aspic", jk = FALSE, copyExe = FALSE)
}
\arguments{
\item{object}{either \emph{biodyn} or \emph{aspic} class}

\item{index}{with relative abundance, \emph{FLQuant} or \emph{FLQuants}, if \strong{object} is of type \emph{bodyn}}

\item{...}{any other parameter}

\item{object;}{an \code{aspic} object}

\item{dir;}{an optional \code{dir} where aspic text files used for fitting can be found}
}
\value{
An aspic object with fitted values and parameter estimates
}
\description{
A generic method for fitting catch and index of relative abundance for both \emph{biodyn} and \emph{aspic}.

Fits the aspic model to catch and catch per unit effort data
}
\seealso{
\code{\link{aspic}}, \code{\link{biodyn}}, \code{\link{jk}}, \code{\link{boot}}

\code{\link{aspic},\link{biodyn},\link{boot},\link{jk}}
}

