% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuantJKs.R
\docType{class}
\name{FLQuantJKs}
\alias{FLQuantJKs}
\alias{FLQuantJKs,ANY-method}
\alias{FLQuantJKs,FLQuantJKs-method}
\alias{FLQuantJKs,list-method}
\alias{FLQuantJKs,missing-method}
\alias{FLQuantJKs-class}
\alias{FLQuantJKs-methods}
\title{Class FLQuantJKs}
\description{
\code{FLQuantJKs} is a \code{list} of \code{FLQuant} objects.
It is very similar to the standard \code{list} class.
It implements a lock mechanism that, when turned on, does
not allow the user to increase or decrease the object length.
The elements of \code{FLQuantJKs} must all be of class  \code{FLQuant}.
}
\section{Slots}{
 \describe{
    \item{.Data}{The data. \code{list}.}
    \item{names}{Names of the list elements. \code{character}.}
    \item{desc}{Description of the object. \code{character}.}
    \item{lock}{Lock mechanism, if turned on the length of the list can not be modified by adding or removing elements. \code{logical}.}
}
}
\keyword{classes}

