% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biodyn-msy.R
\name{fmsy}
\alias{fmsy}
\alias{fmsy,biodyn,missing-method}
\alias{fmsy,biodyn,numeric-method}
\alias{fmsy,character,FLPar-method}
\alias{fmsy,factor,FLPar-method}
\alias{fmsy,fmsy-method}
\title{fmsy}
\usage{
fmsy(object, params, ...)
}
\arguments{
\item{object}{an object of class \code{biodyn} or a string or factor that species the model}

\item{params}{an \code{FLPar} object with model parameters}

\item{...}{any other parameters}
}
\value{
an \code{FLPar} object with value(s) of FMSY
}
\description{
Calculates FMSY given the model parameters, can be done for a biodyn class, or by specifying the model and parameters
}
\examples{
\dontrun{
fmsy('logistic',FLPar(msy=100,k=500))
}
}
\seealso{
\code{\link{msy}}, \code{\link{bmsy}}
}

