% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspic-profile.R
\docType{methods}
\name{fnProfile}
\alias{fnProfile}
\title{fnProfile}
\usage{
fnProfile(x)
}
\arguments{
\item{x:}{an \code{aspic} object}
}
\value{
a \code{data frame} with results by data component.
}
\description{
Outputs as a data.frame a summary of parameters and RSS etc by data component 
generated when doing a profile
}
\examples{
\dontrun{
data(asp)
dcK=profile(asp,which=c("k"),range=seq(0.2,2.0,length.out=21),fn=fnProfile)
ggplot(dcK)+geom_line(aes(k,value,group=variable,col=variable))+
 theme_ms(12,legend.position="bottom")+
 ylab("Residual Sum of Squares")+xlab("K")

}       
}
\seealso{
\code{\link{biodyn},\link{profile} \link{fit}}
}
