% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biodyn-methods.R
\name{harvest}
\alias{harvest}
\alias{harvest,aspic-method}
\title{harvest}
\usage{
harvest(object, ...)
}
\arguments{
\item{object}{either \emph{biodyn} or \emph{aspic} classes}
}
\description{
harvest rate
}

