% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-hcr.R
\name{hcrParam}
\alias{hcrParam}
\title{hcrParam}
\usage{
hcrParam(ftar, btrig, fmin, blim)
}
\arguments{
\item{ftar}{an object of class \code{FLPar}}

\item{btrig}{an object of class \code{FLPar}}

\item{fmin}{an object of class \code{FLPar}}

\item{blim}{an object of class \code{FLPar}}
}
\description{
Combines reference points into the HCR breakpts
}
\examples{
\dontrun{
tac('logistic',FLPar(msy=100,k=500))
}
}
\seealso{
\code{\link{hcr}}
}

