% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-hcr.R
\name{hcrPlot}
\alias{hcrPlot}
\alias{hcrPlot,biodyn-method}
\alias{hcrPlot-method}
\title{hcrPlot}
\usage{
hcrPlot(object, ...)
}
\arguments{
\item{object}{an object of class \code{biodyn} or}

\item{params}{\code{FLPar} object with hockey stock HCR parameters}

\item{maxB}{=1}

\item{rel}{=TRUE}
}
\value{
a \code{FLPar} object with value(s) for HCR
}
\description{
Calculates break pointts for a hockey stick HCR
}
\examples{
\dontrun{
simBiodyn()
}

}
\seealso{
\code{\link{hcr}},  \code{\link{msy}},  \code{\link{bmsy}}, \code{\link{fmsy}}
}

