% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hrate.R
\name{hrate}
\alias{hrate}
\alias{hrate,aspic-method}
\alias{hrate,biodyn-method}
\title{hrate}
\usage{
hrate(object)
}
\arguments{
\item{object}{either \emph{biodyn} or \emph{aspic} classes}
}
\description{
harvest rate
}

