% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspic-constructors.R
\name{is.aspic}
\alias{is.aspic}
\title{is.aspic}
\usage{
is.aspic(x)
}
\arguments{
\item{x}{biodyn class}
}
\value{
TRUE or FALSE
}
\description{
Checks class type and returns TRUE if object is of type biodyn
}
\examples{
\dontrun{
 is.aspic(aspic()) 
 }
}
