% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-constructors.R
\name{is.biodyn}
\alias{is.biodyn}
\title{is.biodyn}
\usage{
is.biodyn(x)
}
\arguments{
\item{x}{biodyn class}
}
\value{
TRUE or FALSE
}
\description{
Checks class type and returns TRUE if object is of type biodyn
}
\examples{
\dontrun{
 is.biodyn(biodyn()) 
 }
}
