% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{jk}
\alias{jk}
\alias{jk,aspic-method}
\alias{jk,biodyn,FLQuant-method}
\alias{jk,biodyn,FLQuants-method}
\title{jk}
\usage{
jk(object, index, ...)
}
\arguments{
\item{object}{either \emph{biodyn} or \emph{aspic} class}

\item{index}{with relative abundance of type \emph{FLQuantJK} or \emph{FLQuantJKs}, if object is of type \emph{biodyn}}

\item{...}{any other parameter}
}
\description{
A generic method for jack knifing a fit to catch and index of relative abundance for both \emph{biodyn} and \emph{aspic}.
}
\seealso{
\code{\link{aspic}}, \code{\link{biodyn}}, \code{\link{fit}}, \code{\link{boot}}
}

