% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-kobe.R
\name{kobe}
\alias{kobe}
\alias{kobe,biodyn,missing-method}
\alias{kobe,biodyns,ANY-method}
\alias{kobe,biodyns,missing-method}
\title{kobe}
\usage{
kobe(object, method, ...)
}
\arguments{
\item{object}{biodyn object}

\item{method}{missing}

\item{...}{other arguments}
}
\value{
data.frame or list of data.frames
}
\description{
Creates time series of stock relative to BMSY and harvest rate relative
to FMSY
}
\examples{
\dontrun{
sim()
}
}

