% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\docType{methods}
\name{mixnorm}
\alias{mixnorm}
\title{mixnorm}
\usage{
mixnorm(n, mn, sd, bin, left = T)
}
\arguments{
\item{mn}{an \code{FLQuant} with means}

\item{sd}{an \code{FLQuant} with standard deviation}
}
\value{
an \code{FLPar} with expected probability for lengths-at-age
}
\description{
Simulates a normal density function from an \code{FLQuant}
}
\examples{
\dontrun{
mn=FLQuant( 1:5,     dimnames=list(age=1:5,year=1991:2000,unit=1:2))
sd=FLQuant((1:5)/5,dimnames=list(age=1:5,year=1991:2000))
md=mixnorm(mn,sd,(1:20)/4)
ggplot(as.data.frame(md))+
   geom_line(aes(len,data,group=age))+
   facet_wrap(year~unit)
ggplot(as.data.frame(apply(md,1,sum)))+
   geom_line(aes(as.numeric(as.character(params)),data))     
 }
}

