% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lk-methods.R, R/resample.R
\docType{methods}
\name{resample,FLQuant-method}
\alias{resample,FLQuant-method}
\alias{resample,FLPar-method}
\title{resample}
\usage{
\S4method{resample}{FLQuant}(x, dim = 2:6, size, replace = T, ...)

\S4method{resample}{FLQuant}(x, dim = 2:6, size, replace = T, ...)
}
\arguments{
\item{dim}{dimensions of \code{FLQuant} to sample within, i.e. if equal to \code{2} then sampling is within \code{year}}

\item{size}{sum of frequency disribution}

\item{replace}{sample with replacement?, defaults to \code{FALSE} only useful if values in \code{object} are integers}

\item{...}{any other parameters}

\item{object}{an \code{FLQuant}}
}
\value{
a \code{FLQuant} with simulated frequency distribution

a \code{FLQuant} with simulated frequency distribution
}
\description{
Resamples from a distribution to generate a frequency table

Resamples from a distribution to generate a frequency table
}
\examples{
\dontrun{
  library(FLCore)
  data(ple4sex)
  
  apply(resample(catch.n(ple4sex)[,1:4],2,  100),2,  sum)
  apply(resample(catch.n(ple4sex)[,1:4],2:3,100),2:3,sum)
  apply(resample(catch.n(ple4sex)[,1:4],1:2,100),1:2,sum)
  }
\dontrun{
  library(FLCore)
  data(ple4sex)
  
  apply(resample(catch.n(ple4sex)[,1:4],2,  100),2,  sum)
  apply(resample(catch.n(ple4sex)[,1:4],2:3,100),2:3,sum)
  apply(resample(catch.n(ple4sex)[,1:4],1:2,100),1:2,sum)
  }
}
