% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lk-mfcl.R
\name{mfclGrw}
\alias{mfclGrw}
\title{Returns von Bertalannfy growth parameters from Multifan-CL}
\usage{
mfclGrw(parfile = getoutputparfile("plot.rep"))
}
\arguments{
\item{parfile}{the par file output from Multifan-cl}
}
\value{
a numeric vector
}
\description{
Returns von Bertalannfy growth parameters from Multifan-CL
}

