% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpmiExtraction.R
\name{mpmiExcraction}
\alias{mpmiExcraction}
\title{Extract words based on MPMI+BE alogrithm}
\usage{
mpmiExcraction(
  df,
  ChineseCharFilter = T,
  lambda = 0.25,
  optics,
  steps = 100,
  threshold,
  bayesianCutoff
)
}
\arguments{
\item{df}{One column dataframe.}

\item{ChineseCharFilter}{Drop non-Chinese characters. Default is True.}

\item{lambda}{Weight distribution between PMI and BE, which will directly affect the results.}

\item{optics}{Automatically optimize the result, select TURE if you don't know how to choose lambda.}

\item{steps}{Number of iterations in the optimization process.}

\item{threshold}{threshold for the final score.}

\item{bayesianCutoff}{threshold for the bayesian probabilty value in word extension process}
}
\value{
List of extracted strings.
}
\description{
Extract words based on MPMI+BE alogrithm
}
\examples{
\donttest{
mpmiExcraction(as.data.frame(drugnames[1:20,]),
               ChineseCharFilter = T,
               lambda = 0.25,
               optics = T,
               steps = 100,
               threshold = 30,
               bayesianCutoff = 0.6)
}
}
