% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-mvn.R
\name{mvn}
\alias{mvn}
\title{mvn}
\usage{
mvn(object, n, nms = dimnames(object@control[object@control[, "phase", ]
  > 0, ])$params, fwd = FALSE, params = TRUE)
}
\arguments{
\item{object}{a \code{biodyn} object}

\item{n}{\code{numeric} with number of \code{iter} to create}

\item{nms}{the parameters for Monte Carlo simulation}

\item{fwd}{\code{logical} do you want to simuate historic time series of stock biomass?, default is \code{FALSE}}

\item{params}{actual parameters}
}
\value{
\code{biodyn} with simuated time series
}
\description{
Takes a fitted biodyn object and uses the covariance matrix (in the \code{vcov} slot)  and the
parameter estimates (\code{params} slot) to use Monte Carlo simulation to generate new 
parameters
}
\examples{
\dontrun{
x=1
}
}
