% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextWordMiner.R
\name{nextWordMiner}
\alias{nextWordMiner}
\title{Mining next word}
\usage{
nextWordMiner(n_char_final, numOfChar, bayesianCutoff)
}
\arguments{
\item{n_char_final}{word table with n characters}

\item{numOfChar}{The number of character of candidate word, which should be equal to the number of characters in n_char_final}

\item{bayesianCutoff}{Bayesian conditional probability of occurrence of next word. The algorithm will drop}
}
\value{
List of new candidates words and words which cannot be extended.
}
\description{
Extend candidate words based on Bayesian probability.
}
\examples{
\donttest{
nextWordMiner(df, 2, 0.3)
}
}
