% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-nll.R
\docType{methods}
\name{nll,FLQuant,FLQuant,FLPar-method}
\alias{nll,FLQuant,FLQuant,FLPar-method}
\alias{nll,FLQuant,FLQuants,FLPar-method}
\alias{nll,biodyn,FLQuant,missing-method}
\alias{nll,biodyn,FLQuants,missing-method}
\title{nll}
\usage{
\S4method{nll}{FLQuant,FLQuant,FLPar}(object, index, params, min = 0.01)
}
\arguments{
\item{object}{an \code{FLQuant} with a time series of catch, iters must be equal to 1}

\item{params}{an \code{FLPar} object with parameters for the production function, r, k, p and b0.}

\item{min, }{the minimum permissable population level, used to check that the catch can be taken.}

\item{...}{any other parameters}
}
\value{
a \code{FLPar} a subset of params with parameter values that can explain the catch
}
\description{
Checks that the parameters can be taken given the catch time series
}
\examples{
\dontrun{
params=nllFn(catch,params)
}
}
