% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmc.R
\name{nmcFn}
\alias{nmcFn}
\title{nmc}
\usage{
nmcFn(x, s, E = 5, z = 1.96)
}
\arguments{
\item{x}{\code{FLQuant}}

\item{E}{\code{numeric}  The required percentage error of the mean}

\item{z}{\code{numeric}  Value of Z for a given confidence level for a normally distributed random variable, default is 1.96 for a 95% CI}

\item{...}{other arguments}
}
\value{
\code{FLQuant} with number of iters required by number of iters
}
\description{
Calculates the number of iterations needed in Monte Carlo Simulation to achieve a given precision.
}
\examples{
 \dontrun{
 data(ple4) 
 ssb=rlnorm(2000,log(ssb(ple4)),0.5)
 nmc(ssb) 
 }
 
}
