% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-oem.R
\docType{methods}
\name{oem,FLStock-method}
\alias{oem,FLStock-method}
\title{oem}
\usage{
\S4method{oem}{FLStock}(object, cv = rlnorm(dim(stock(object))[6], FLQuant(0,
  dimnames = dimnames(stock(object))[-6]), 0.3), timing = 0.5, mult = TRUE,
  fishDepend = FALSE, effort = c("f", "h"), mass = TRUE,
  q = FLQuant(cumprod(1 + rep(0, dim(fbar(object))[2])), dimnames =
  dimnames(fbar(object))), sel = FLQuant(FLQuant(1, dimnames =
  dimnames(harvest(object)))), seed = NULL)
}
\arguments{
\item{object}{\code{FLStock} for which an index is to be generated}

\item{cv}{either a \code{numeric} with a CV on the logscale or an \code{FLQuant} 
with random deviates}

\item{mult}{\code{logical} that determines whether the deviates are multiplicative or}

\item{fishDepend}{\code{logical} that determines whether the index is proportional to the 
stock or catch/effort, by default set to \code{TRUE}}

\item{effort}{\code{character} c("h","f") that determines what proxy to use for effort,
i.e. harvest rate or instanteous fishing mortality}

\item{mass}{\code{logical} default is TRUE, whether index is in mass or numbers}

\item{q}{\code{numeric} or \code{FLQuant} that specifies change in catchability}

\item{sel}{\code{FLQuant} vector at age that shapes the catch or biomass.
additive, by default set to \code{TRUE},}

\item{...}{other arguments}
}
\value{
\code{FLQuant} with time series
}
\description{
Creates an \code{FLQuant} to represent an index of relative abundance
}
\examples{
\dontrun{
 data(ple4) 
 cpue=oem(ple4) 
 }
}

