% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{om}
\alias{om}
\title{om}
\format{
An \code{FLStock} object
\describe{
  \item{stock.n}{numbers-at.age, \code{FLQuant}}
  \item{stock.wt}{mass-at-age, \code{FLQuant}}
  ...
}

An \code{FLBRP} object
\describe{
  \item{stock.wt}{mass-at-age, \code{FLQuant}}
  ...
}
}
\description{
An example Operating Model, based on North Atlantic Albacore

Reference points and expected dynamics, for Operating Model based on North Atlantic Albacore
}
\details{
A dataset containing the historic stock estimates, derived from the ICCAT multifan-CL assessment

A dataset containing the average estimates of biological parameters and selection patterns, 
derived from the ICCAT multifan-CL assessment
}
\references{
\url{http://http://iccat.int/Documents/CVSP/CV070_2014/n_4/CV070041365.pdf/}

\url{http://http://iccat.int/Documents/CVSP/CV070_2014/n_4/CV070041365.pdf/}
}
\author{
Laurence Kell \email{laurie@kell.es}

Laurence Kell \email{laurie@kell.es}
}
\keyword{data}
