% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-plot.R
\docType{methods}
\name{plot,biodyn,missing-method}
\alias{plot,biodyn,missing-method}
\alias{plot,biodyns,missing-method}
\alias{plotCcf,FLQuants-method}
\alias{plotCcf,biodyns,missing-method}
\title{plot}
\usage{
\S4method{plot}{biodyn,missing}(x, y, probs = c(0.9, 0.75, 0.5, 0.25,
  0.1), na.rm = FALSE, type = 7, worm = NULL, fn = list(Stock =
  function(x) stock(x), Harvest = function(x) harvest(x), Yield =
  function(x) catch(x)), facet = facet_wrap(~qname, scales = "free", ncol
  = 1), ...)

\S4method{plotCcf}{FLQuants}(data, ...)
}
\arguments{
\item{x}{an object of class \code{biodyn}}

\item{y}{second argument}

\item{probs}{numeric vector of probabilities with values in [0,1].}

\item{na.rm}{a logical value indicating whether NA values should be stripped before computation.}

\item{type}{an integer between 1 and 9 selecting one of the quantile algorithms to be used.}

\item{worm}{iter(s) to plot as lines}

\item{fn}{functions}

\item{facet}{facet for panelling}

\item{x}{an object of class \code{biodyn}}
}
\value{
an \code{ggplot2} object

an \code{ggplot2} object
}
\description{
Plots time series of biomass, harvest rate and catch for a \code{biodyn} object, using \code{ggplot2}.


}
\examples{
\dontrun{

bd =sim()
plot(bd)
} 
\dontrun{

} 
}
