% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-plot.R
\docType{methods}
\name{plotEql,biodyn,missing-method}
\alias{plotEql,biodyn,missing-method}
\title{plotEql}
\usage{
\S4method{plotEql}{biodyn,missing}(data, biomass, ...)
}
\arguments{
\item{biomass}{an object of holding biomass at beginning of year}

\item{...}{other arguments}

\item{x}{an object of class \code{biodyn}}
}
\value{
an \code{ggplot2} object
}
\description{
Creates a \code{ggplot2} object that plots time series of biomass, harvest rate and catch. The basic object can then be modified by adding ggpot2 layers.
}
\examples{
\dontrun{
 refpts('logistic',FLPar(msy=100,k=500))
}
}
\seealso{
\code{\link{plotProduction}}\code{\link{plotMSE}}
}

