% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-plot.R
\docType{methods}
\name{plotHcr,biodyn-method}
\alias{plotHcr,biodyn-method}
\alias{plotHcr-method}
\title{plotHcr}
\usage{
\S4method{plotHcr}{biodyn}(object, params = FLPar(ftar = 0.7, btrig =
  0.7, fmin = 0.01, blim = 0.2), maxB = 1, rel = TRUE)
}
\arguments{
\item{object}{an object of class \code{biodyn} or}

\item{params}{\code{FLPar} object with hockey stock HCR parameters}

\item{maxB}{=1}

\item{rel}{=TRUE}
}
\value{
a \code{FLPar} object with value(s) for HCR
}
\description{
Plots a hockey stick HCR with break pointts
}
\examples{
\dontrun{
simBiodyn()
}

}
