% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-plot.R
\docType{methods}
\name{plotIndex,FLQuants-method}
\alias{plotIndex,FLQuants-method}
\alias{plotIndex,biodyns,missing-method}
\title{plotIndex}
\usage{
\S4method{plotIndex}{FLQuants}(data, facet = facet_wrap(~qname, ncol = 1,
  scales = "free_y"), ...)
}
\arguments{
\item{x}{an object of class \code{biodyn}}
}
\value{
an \code{ggplot2} object
}
\description{

}
\examples{
\dontrun{

} 
}

