% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndex.R
\name{plotIndexResidual,FLQuants-method}
\alias{plotIndexResidual,FLQuants-method}
\alias{plotIndexResidual,biodyn,missing-method}
\alias{plotIndexResidual,aspic,missing-method}
\alias{plotIndexResidual,FLQuants,missing-method}
\title{plotIndexResidual}
\usage{
\S4method{plotIndexResidual}{FLQuants}(data, ...)
}
\arguments{
\item{data}{an object of class \code{FLQuants} \code{biodyn} \code{aspic}}
}
\value{
an \code{ggplot2} object
}
\description{

}
\examples{
\dontrun{

} 
}
