% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBiodyn.R
\name{plotJack}
\alias{plotJack}
\title{plotJack}
\usage{
plotJack(x, y, ncol = 1)
}
\arguments{
\item{x}{an object of class \code{biodyn} that has been jack knifed, i.e. by 
providing a jack knifed CPUE series to fit}

\item{y}{the original \code{biodyn} object}

\item{ncol}{number of colums in plot panel}
}
\value{
an \code{ggplot2} object
}
\description{
Create a \code{ggplot2} plot based on a jack knifed biodyn and plots 
time series of biomass and harvest rate. 
The basic object can then be modified by adding ggpot2 layers.
}
\examples{
\dontrun{
#simulate an object with known properties
bd=sim()
bd=window(bd,end=49)

#simulate a proxy for stock abundance
cpue=(stock(bd)[,-dims(bd)$year]+stock(bd)[,-1])/2
cpue=rlnorm(1,log(cpue),.2)

#set parameters
setParams(bd) =cpue
setControl(bd)=params(bd)
control(bd)[3:4,"phase"]=-1

#fit
bd=fit(bd,cpue)

bdJK=fit(bd,jackknife(cpue))
plotJack(bdJK)
bd  =randJack(100,bd)
}
}

