% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBiodyn.R
\docType{methods}
\name{plotPrd,biodyn,missing-method}
\alias{plotPrd,biodyn,FLBRP-method}
\alias{plotPrd,biodyn,FLQuant-method}
\alias{plotPrd,biodyn,missing-method}
\title{plotPrd}
\usage{
\S4method{plotPrd}{biodyn,missing}(data, biomass, ...)
}
\arguments{
\item{biomass}{optional argument, an FLQuant with biomass at beginning of year}

\item{...}{other arguments}

\item{x}{an object of class \code{biodyn}}
}
\value{
an \code{ggplot2} object
}
\description{
Creates a \code{ggplot2} object that plots equilibrium values of biomass, harvest rate and catch against each other.
The basic object can then be modified by adding ggpot2 layers.
}
\examples{
\dontrun{
 refpts('logistic',FLPar(msy=100,k=500))
}
}
\seealso{
\code{\link{plotMSE}}, \code{\link{plotEql}}
}

