% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspic-plot.R, R/biodyn-plot.R
\name{plotProduction}
\alias{plotProduction}
\alias{plotProduction,biodyn,missing-method}
\alias{plotProduction,biodyn,FLBRP-method}
\alias{plotProduction,biodyn,FLQuant-method}
\alias{plotProduction,biodyns,missing-method}
\title{plot aspics,missing
plotIndex aspic
plotIndex aspics}
\usage{
plotProduction(data, biomass, ...)

\S4method{plotProduction}{biodyn,missing}(data, biomass, ...)
}
\arguments{
\item{biomass}{optional argument, an FLQuant with biomass at beginning of year}

\item{...}{other arguments}

\item{object}{an object of class \code{biodyn}}
}
\value{
an \code{ggplot2} object
}
\description{
Creates a \code{ggplot2} object that plots equilibrium values of biomass, harvest rate and catch against each other.
The basic object can then be modified by adding ggpot2 layers.
}
\details{
plotProduction
}
\examples{
\dontrun{
 refpts('logistic',FLPar(msy=100,k=500))
}
}
