% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/biodyn-methods.R
\docType{methods}
\name{production}
\alias{production}
\alias{production,FLBRP,missing-method}
\alias{production,biodyn,missing-method}
\alias{production,biodyn-method}
\title{production}
\usage{
production(object, biomass, ...)

\S4method{production}{biodyn,missing}(object, biomass = stock(object))
}
\arguments{
\item{object}{either \emph{biodyn} or \emph{FLBRP} class}

\item{biomass}{}

\item{...}{any other parameter}

\item{object}{an object of class \code{biodyn}}

\item{biomass}{stock biomaas, may be a \code{numerix},  \code{FLQuant} or missing. In the latte case the stock slot will be used.}

\item{...}{other arguments}
}
\value{
an \code{FLPar} object
}
\description{
Estimates production for a given biomass

Calculates the surplus production for a biomass dynamic model given a level of stock biomass
}
\examples{
\dontrun{ production(bd,seq(0,params(bd)['k'])) }
 
}
\seealso{
\code{\link{plotProduction}}
}

