% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-rand.R
\name{randFn}
\alias{rand,numeric,FLPar,FLPar-method}
\alias{rand,numeric,FLQuant,FLQuant-method}
\alias{rand,numeric,biodyn,biodyn-method}
\alias{rand-method}
\alias{randFn}
\title{rand}
\usage{
randFn(n, object, sim)
}
\arguments{
\item{n}{\code{numeric} with number of simulations}

\item{object}{\code{biodyn}}

\item{sim}{a \code{biodyn} object}

\item{...}{other arguments}
}
\value{
\code{biodyn} with estimates of stock based on catch time series
}
\description{
Simulates a \code{biodyn} object for a catch series, given the parameter estimates  in the \code{param} 
slot and variance covariance matrix 
http://young.physics.ucsc.edu/jackboot.pdf
}
\examples{
\dontrun{
#simulate an object with known properties
bd=sim()
bd=window(bd,end=49)

#simulate a proxy for stock abundance
cpue=(stock(bd)[,-dims(bd)$year]+stock(bd)[,-1])/2
cpue=rlnorm(1,log(cpue),.2)

#set parameters
setParams(bd) =cpue
setControl(bd)=params(bd)
control(bd)[3:4,"phase"]=-1

#fit
bd=fit(bd,cpue)

bdJK=fit(bd,jackknife(cpue))


bdRnd  =rand(100,bd,bdJK)

plot(rand(100,stock(bd)[,40:45],stock(bdJK)[,40:45]))
}
}

