% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{readAspic}
\alias{readAspic}
\alias{readAspic,character-method}
\title{readAspic}
\usage{
readAspic(object, ...)
}
\arguments{
\item{object}{an \strong{file path}}

\item{...}{any other parameter}
}
\description{
Read ASPIC text files, either \code{inp} for inputs, or output files, produced by the executable version of ASPIC
}
\examples{
\dontrun{
    readAspic("aspic.inp")}

}
\seealso{
\code{\link{readFLStock}}, \code{\link{writeAspic}}
}

