% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biodyn-msy.R
\name{refpts}
\alias{refpts}
\alias{refpts,ANY}
\alias{refpts,biodyn,missing-method}
\alias{refpts,character,FLPar-method}
\alias{refpts,factor,FLPar-method}
\alias{refpts-method}
\title{Maximum Sustainable Yield reference points}
\usage{
refpts(object, params, ...)
}
\arguments{
\item{object}{an object of class \code{biodyn} or a string or factor that species the model}

\item{params}{an \code{FLPar} object with model parameters}

\item{...}{any other arguments}
}
\value{
an \code{FLPar} object with value(s) of F_MSY
}
\description{
Calculates MSY, B_MSY and F_MSY given the model parameters, can be done for a biodyn class, or by specifying the model and parameters
}
\examples{
\dontrun{
refpts('logistic',FLPar(msy=100,k=500))
}
}
\seealso{
\code{\link{msy}}, \code{\link{bmsy}}
}

