% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{setParams<-}
\alias{setControl<-,biodyn,FLQuant-method}
\alias{setControl<-,biodyn,FLQuants-method}
\alias{setParams<-}
\title{setParams<-}
\usage{
setParams(object, ...) <- value
}
\arguments{
\item{object}{\code{biodyn}}

\item{value}{CPUE as \code{FLQuant} or \code{FLQuants}}

\item{...}{any other parameter}
}
\description{
Sets catchability \code{q} and CV \code{sigma} in the params \code{FLPar} slot 
for CPUE provided either an \code{FLQuant} or \code{FLQuants}
}
\examples{
\dontrun{
setParams(bd)=cpue
}
 
}

