% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biodyn-oem.R
\name{sim}
\alias{sim}
\alias{sim,FLStock,ANY-method}
\alias{sim,missing,missing-method}
\alias{sim-method}
\title{sim}
\usage{
sim(stock, brp, ...)
}
\arguments{
\item{...}{other arguments}

\item{model}{character corresponding to model}

\item{params}{surplus production parameters}

\item{harvest}{\code{FLQuant} with harvest rate}

\item{bounds}{on \code{control}}
}
\value{
biodyn object with simulated time series
}
\description{
Creates a biodyn object with known properties
}
\examples{
\dontrun{
 bd=sim()
 }
}

