% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-methods.R
\docType{methods}
\name{production,biodyn,missing-method}
\alias{production,FLBRP,FLStock-method}
\alias{production,FLStock,missing-method}
\alias{production,biodyn,FLQuant-method}
\alias{production,biodyn,missing-method}
\alias{production,biodyn,numeric-method}
\title{production}
\usage{
\S4method{production}{biodyn,missing}(object, biomass = stock(object))
}
\arguments{
\item{object}{an object of class \code{biodyn}}

\item{biomass}{stock biomaas, may be a \code{numerix},  \code{FLQuant} or missing. In the latte case the stock slot will be used.}

\item{...}{other arguments}
}
\value{
an \code{FLPar} object
}
\description{
Calculates the surplus production for a biomass dynamic model given a level of stock biomass
}
\examples{
\dontrun{ production(bd,seq(0,params(bd)['k'])) }
 
}
\seealso{
\code{\link{plotProduction}}
}

