% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mng-methods.R
\name{fapexAge}
\alias{fapexAge}
\title{mng}
\usage{
fapexAge(object)
}
\arguments{
\item{object}{either  \emph{FLStock} \emph{biodyn} or \emph{aspic} classes}
}
\description{
calculates time series of quantities useful for management
}

