% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mng-methods.R
\docType{methods}
\name{timeSeries,FLStock,FLPar-method}
\alias{timeSeries,FLStock,FLPar-method}
\alias{ts,FLStock,FLBRP-method}
\alias{ts,FLStock,FLPar-method}
\alias{ts,FLStocks,FLBRP-method}
\alias{ts,FLStocks,FLBRPs-method}
\alias{ts,biodyn,FLPar-method}
\title{ts}
\usage{
\S4method{timeSeries}{FLStock,FLPar}(object, params, df = TRUE)
}
\arguments{
\item{object}{either  \emph{FLStock} \emph{biodyn} or \emph{aspic} classes}
}
\description{
calculates time series of quantities useful for management
}
