% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mng-methods.R
\name{ts}
\alias{mng}
\alias{ts}
\title{ts}
\usage{
ts(object, params, ...)

mng(object, params, ...)
}
\arguments{
\item{object}{either  \emph{FLStock} \emph{biodyn} or \emph{aspic} classes}

\item{object}{either  \emph{FLStock} \emph{biodyn} or \emph{aspic} classes}
}
\description{
calculates time series of quantities useful for management

calculates time series of quantities useful for management
}

