% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{two_char}
\alias{two_char}
\title{Frequency, PMI and BE dataframe of Chinese words wtih two characters.}
\format{A data frame with 1 variables:
\describe{
\item{\code{character}}{chinese word with two characters.}
\item{\code{freq}}{Occurence times of target word in real world.}
\item{\code{pct}}{Frequency of target word in real world.}
\item{\code{A}}{entropy of first character in target word.}
\item{\code{B}}{entropy of second character in target word.}
\item{\code{real_log}}{pmi value of target word.}
}}
\usage{
two_char
}
\description{
This dataframe contains Chinese Word with two characters with its branch entropy and pmi value
}
\examples{
\dontrun{
 two_char
 }
}
\keyword{datasets}
