% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userDataImport.R
\name{userDataImport}
\alias{userDataImport}
\title{Occurrence and frequency of the next word}
\usage{
userDataImport(df, ChineseCharFilter = TRUE)
}
\arguments{
\item{df}{user import file}

\item{ChineseCharFilter}{Drop non-Chinese characters. Default is True.}
}
\value{
A dataframe containing the next word and its occurrence and frequency
}
\description{
Processing user data
}
\examples{
\donttest{
userDataImport(drugnames)
}
}
