% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{writeAspic}
\alias{writeAspic}
\alias{writeAspic,aspic-method}
\title{writeAspic}
\usage{
writeAspic(object, ...)
}
\arguments{
\item{object}{an \strong{apic} object}

\item{...}{any other parameter}
}
\description{
Writes the ASPIC text input file \code{inp} to a file or connection.
The executable version of ASPIC uses an input file, this method generates that file
}
\examples{
\dontrun{
    writeAspic(albn,"aspic.inp")}

}
\seealso{
\code{\link{writeFLStock}}, \code{\link{readAspic}}
}

