% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-xval.R, R/aspic-xval.R
\docType{methods}
\name{xval,biodyn,ANY-method}
\alias{xval,biodyn,ANY-method}
\alias{xval,biodyn,missing-method}
\alias{xval,biodyn-method}
\alias{xval-method}
\title{xval}
\usage{
\S4method{xval}{biodyn,ANY}(object, index, minyear = range(object)["maxyear"]
  - 4, maxyear = range(object)["maxyear"])

\S4method{xval}{biodyn,missing}(object, minyear = range(object)["maxyear"] -
  4, maxyear = range(object)["maxyear"], mf = TRUE)
}
\arguments{
\item{object}{an object of class \code{biodyn}}

\item{index}{an \code{FLQuant} or \code{FLQuants} with index of relative stock abundance}

\item{minyear}{last year to fit data to}

\item{maxyear}{last year to project to, by default is the last year in the catch}

\item{object}{an object of class \code{biodyn}}

\item{index}{an \code{FLQuant} or \code{FLQuants} with index of relative stock abundance}

\item{minyear}{last year to fit data to}

\item{maxyear}{last year to project to, by default is the last year in the catch}
}
\description{
Performs a cross-validation uisng a hindcast

Performs a cross-validation uisng a hindcast
}
\details{
Returns a data.frame with index, year, obs and hat

Returns a data.frame with index, year, obs and hat
}
\examples{
\dontrun{
 data(bd)  
}
\dontrun{
 data(bd)  
}
}

