% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-coerce.R
\docType{methods}
\name{as}
\alias{as}
\title{as("biodyn", "FLStock")}
\usage{
\S4method{instantaneous}{FLQuant}(object)
}
\description{
Coerces an \code{FLStock} into a \code{biodyn} object.
}
\concept{biodyn}
