% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspic-fit.R
\docType{methods}
\name{fit}
\alias{ASPIC}
\alias{Bootstraps}
\alias{biomass}
\alias{boot,}
\alias{chkIters}
\alias{dynamic}
\alias{fit}
\alias{fit,aspic,missing-method}
\alias{model.}
\alias{the}
\title{fit}
\usage{
chkIters(object)
}
\arguments{
\item{object;}{an \code{aspic} object}

\item{dir;}{an optional \code{dir} where aspic text files used for fitting can be found}

\item{object;}{an \code{aspic} object or}

\item{object;}{a character string giving an aspic "inp" file}

\item{dir;}{an optional \code{dir} where aspic text files used for fitting can be found}

\item{object;}{an \code{aspic} object or}

\item{object;}{a character string giving an aspic "inp" file}

\item{dir;}{an optional \code{dir} where aspic text files used for fitting can be found}
}
\value{
An aspic object with fitted values and parameter estimates

An aspic object with fitted values and parameter estimates

An aspic object with fitted values and parameter estimates
}
\description{
Fits the aspic model to catch and catch per unit effort data

Bootstraps the aspic model

Fits the aspic model to catch and catch per unit effort data removing 1 cpue observation at a time
}
\examples{
\dontrun{
    data(asp)
    asp=boot(asp)}
jk

}
\seealso{
\code{\link{aspic},\link{biodyn},\link{boot},\link{jk}}

\code{\link{biodyn},\link{boot},\link{jk}}

\code{\link{biodyn},\link{boot},\link{fit}}
}

