% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspics.R
\docType{class}
\name{aspics-class}
\alias{aspics-class}
\title{aspics list}
\value{
aspics object
}
\description{
A class that contains a list of ASPIC biomass dynamic stock assessment model classes
}
\examples{
\dontrun{aspics()}

}

