% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyns.R
\docType{data}
\name{biodyns}
\alias{biodyns}
\title{biodyns}
\format{An object of class \code{NULL} of length 0.}
\arguments{
\item{object}{can be \code{biodyn} object or a \code{list} of \code{biodyn} objects}

\item{...}{additional \code{biodyn} objects}
}
\value{
\code{biodyns} object
}
\description{
Create a list of biodyn objects

#'  biodyns-method biodyns,biodyn-method  biodyns,list-method  biodyns,missing-method biodyns,aspics-method
}
\examples{
\dontrun{
   biodyns(biodyn())
   }
}
\keyword{datasets}

