% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-boot.R
\docType{methods}
\name{boot,biodyn-method}
\alias{boot,biodyn-method}
\alias{boot-method}
\title{boot, Bootstraps biodyn.}
\usage{
\S4method{boot}{biodyn}(object, run = TRUE)
}
\arguments{
\item{object;}{a \code{biodyn} object}
}
\description{
Bootstraps the aspic model
}
\examples{
\dontrun{
    data(asp)
    asp=boot(asp)}
}
\seealso{
\code{\link{biodyn},\link{boot},\link{jk}}
}

