% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-accessors.R
\docType{methods}
\name{control,biodyn-method}
\alias{control,biodyn-method}
\title{control}
\usage{
\S4method{control}{biodyn}(object)
}
\arguments{
\item{object}{\code{biodyn} object}
}
\value{
FLPar
}
\description{
sets initial guess and lower and upper bounds
}
\examples{
\dontrun{control(biodyn())}
}

