% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biodyn-accessors.R, R/generics.R
\name{control}
\alias{control}
\alias{control,biodyn,missing-method}
\alias{control,biodyn-method}
\alias{control-method}
\alias{control<-,biodyn,FLPar-method}
\alias{setControl<-}
\alias{setControl<-,biodyn,FLPar-method}
\alias{setControl<-,biodyn,FLQuant-method}
\alias{setControl<-,biodyn,FLQuants-method}
\alias{setParams<-}
\alias{setParams<-,aspic,data.frame-method}
\alias{setParams<-,biodyn,FLBRP-method}
\alias{setParams<-,biodyn,FLPar-method}
\alias{setParams<-,biodyn,FLQuant-method}
\alias{setParams<-,biodyn,FLQuants-method}
\alias{setParams<-,biodyn,data.frame-method}
\title{control}
\usage{
control(object, ...)

setControl(object, ...) <- value

setParams(object, ...) <- value
}
\arguments{
\item{object}{either \emph{biodyn} or \emph{aspic} class}

\item{...}{any other parameter}

\item{value}{}

\item{object}{either \emph{biodyn} or \emph{aspic} class}

\item{value}{}

\item{...}{any other parameter}
}
\value{
FLPar
}
\description{
sets initial guess and lower and upper bounds

Slot for fitting options, i.e. for starting guesses, to fix parameters or set bounds

Slot for estimates
}
\examples{
\dontrun{control(biodyn())}
}

