% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{control}
\alias{control}
\alias{control,biodyn,missing-method}
\alias{control-method}
\title{control}
\usage{
control(object, ...)
}
\arguments{
\item{object}{either \emph{biodyn} or \emph{aspic} class}

\item{...}{any other parameter}

\item{value}{}
}
\description{
Slot for fitting options, i.e. for starting guesses, to fix parameters or set bounds
}

