% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{diags}
\alias{control,biodyn,missing-method}
\alias{control-method}
\alias{diags}
\title{control}
\usage{
diags(object)
}
\arguments{
\item{object}{either \emph{biodyn} or \emph{aspic} class}

\item{value}{}

\item{...}{any other parameter}

\item{object}{either \emph{biodyn} or \emph{aspic} class}

\item{value}{}

\item{...}{any other parameter}
}
\description{
Slot for fitting options, i.e. for starting guesses, to fix parameters or set bounds

Slot for estimates
}

