% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aspic-fit.R, R/generics.R
\docType{methods}
\name{fit}
\alias{chkIters}
\alias{fit}
\alias{fit,aspic,missing-method}
\alias{fit,aspic-method}
\alias{fit,aspics-method}
\alias{fit,biodyn,FLQuant-method}
\alias{fit,biodyn,FLQuants-method}
\title{fit}
\usage{
chkIters(object)

fit(object, index, ...)
}
\arguments{
\item{object}{either \emph{biodyn} or \emph{aspic} class}

\item{index}{with relative abundance, \emph{FLQuant} or \emph{FLQuants}, if \strong{object} is of type \emph{bodyn}}

\item{...}{any other parameter}

\item{object;}{an \code{aspic} object}

\item{dir;}{an optional \code{dir} where aspic text files used for fitting can be found}

\item{object;}{an \code{aspic} object or}

\item{object;}{a character string giving an aspic "inp" file}

\item{dir;}{an optional \code{dir} where aspic text files used for fitting can be found}

\item{object;}{an \code{aspic} object or}

\item{object;}{a character string giving an aspic "inp" file}

\item{dir;}{an optional \code{dir} where aspic text files used for fitting can be found}
}
\value{
An aspic object with fitted values and parameter estimates

An aspic object with fitted values and parameter estimates

An aspic object with fitted values and parameter estimates
}
\description{
Fits the aspic model to catch and catch per unit effort data

A generic method for fitting catch and index of relative abundance for both \emph{biodyn} and \emph{aspic}.
}
\examples{
jk
jack knifes \\code{aspic}
}
\seealso{
\code{\link{aspic},\link{biodyn},\link{boot},\link{jk}}

\code{\link{biodyn},\link{boot},\link{fit}}

\code{\link{biodyn},\link{boot},\link{jk}}

\code{\link{aspic}}, \code{\link{biodyn}}, \code{\link{jk}}, \code{\link{boot}}
}

