% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-fit.R
\name{fit}
\alias{fit}
\alias{fit,aspic-method}
\alias{fit,aspics-method}
\alias{fit,biodyn,FLQuant-method}
\alias{fit,biodyn,FLQuants-method}
\title{fit}
\usage{
fit(object, index, ...)
}
\arguments{
\item{object}{either \emph{biodyn} or \emph{aspic} class}

\item{index}{with relative abundance, \emph{FLQuant} or \emph{FLQuants}, if \strong{object} is of type \emph{bodyn}}

\item{...}{any other parameter}
}
\description{
A generic method for fitting catch and index of relative abundance for both \emph{biodyn} and \emph{aspic}.
}
\seealso{
\code{\link{aspic}}, \code{\link{biodyn}}, \code{\link{jk}}, \code{\link{boot}}
}

