% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-fit.R
\name{fitPella}
\alias{fitPella}
\title{-mcmc 100000 -mcsave 100'}
\usage{
fitPella(object, dir = tempdir(),
  cmdOps = paste("-maxfn 500 -iprint 0"), lav = FALSE, maxF = 2.5,
  silent = !TRUE)
}
\description{
-mcmc 100000 -mcsave 100'
}
