% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-fwd.R
\docType{methods}
\name{fwd,biodyn,missing-method}
\alias{fwd,FLPar,missing-method}
\alias{fwd,biodyn,FLQuants-method}
\alias{fwd,biodyn,missing-method}
\alias{fwd-method}
\title{fwd}
\usage{
\S4method{fwd}{biodyn,missing}(object, catch = NULL, harvest = NULL,
  stock = NULL, hcr = NULL, pe = NULL, peMult = TRUE, minF = 0,
  maxF = 2, bounds = list(catch = c(Inf, Inf)), lag = 0, end = NULL,
  starvationRations = 0.75, ...)
}
\arguments{
\item{object}{an object of class \code{biodyn} or  \code{biodyns}}

\item{...}{any other parameters}

\item{ctrl}{missing}
}
\description{
Projects a \code{biodyn} object (i.e. a biomass dynamic model)
for a given future catch, harvest rate, or stock biomass. Only
one of these has to be supplied as an argument.
}
\examples{
\dontrun{
bd=simBidyn()
harvest=rlnorm(100,log(harvest(bd))[,-dims(bd)$year],.1)

bdHat =fwd(bdHat,harvest=harvest)

plot(bdHat,worm=c(2,8))+
theme(legend.position="bottom")
 }
}

