% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-methods.R
\name{harvest,biodyn,missing-method}
\alias{harvest,biodyn,missing-method}
\alias{harvest}
\alias{harvest,biodyn-method}
\alias{harvest,aspic-method}
\title{harvest}
\usage{
\S4method{harvest}{biodyn,missing}(object)
}
\arguments{
\item{object}{either \emph{biodyn} or \emph{aspic} classes}
}
\description{
harvest rate
}
