% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-hcr.R, R/FLStock-hcr.R
\name{hcr}
\alias{hcr}
\alias{hcr,biodyn-method}
\title{hcr}
\usage{
hcr(object, refs, ...)

hcr(object, refs, ...)
}
\arguments{
\item{object}{an object of class \code{biodyn} or}

\item{...}{other parameters, i.e.
params \code{FLPar} object with hockey stick HCR parameters, see hcrParam
yr numeric vector with years used to values in HCR
byr numeric vector with years used for bounds
hyr numeric vector with years to use in projection
tac \code{logical} should return value be TAC rather than F?
bndF \code{vector} with bounds (i.e.min and max values) on iter-annual variability on  F
bndTac \code{vector} with bounds (i.e. min and max values) on iter-annual variability on TAC}

\item{object}{an object of class \code{FLStock} or}

\item{...}{other parameters, i.e.
params \code{FLPar} object with hockey stick HCR parameters, see hcrParam
yr numeric vector with years used to values in HCR
byr numeric vector with years used for bounds
hyr numeric vector with years to use in projection
tac \code{logical} should return value be TAC rather than F?
bndF \code{vector} with bounds (i.e.min and max values) on iter-annual variability on  F
bndTac \code{vector} with bounds (i.e. min and max values) on iter-annual variability on TAC}
}
\value{
\code{FLPar} object with value(s) for F or TAC if tac==TRUE

\code{FLPar} object with value(s) for F or TAC if tac==TRUE
}
\description{
Harvest Control Rule, calculates F, or Total Allowable Catch (TAC) based on a hockey stock harvest control rule.

Harvest Control Rule, calculates F, or Total Allowable Catch (TAC) based on a hockey stock harvest control rule.
}
\examples{
\dontrun{
bd   =sim()

bd=window(bd,end=29)
for (i in seq(29,49,1))
bd=fwd(bd,harvest=hcr(bd,yr=i,yr=i+1)$hvt)
}
\dontrun{
bd   =sim()

bd=window(bd,end=29)
for (i in seq(29,49,1))
bd=fwd(bd,harvest=hcr(bd,byr=i,yrs=i+1)$hvt)
}
}
\seealso{
\code{\link{bmsy}}, \code{\link{fmsy}}, \code{\link{fwd}} and \code{\link{hcrParam}}

\code{\link{bmsy}}, \code{\link{fmsy}}, \code{\link{fwd}} and \code{\link{hcrParam}}
}

