% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-methods.R, R/hrate.R
\docType{methods}
\name{harvest,biodyn,missing-method}
\alias{harvest}
\alias{harvest,aspic-method}
\alias{harvest,biodyn,missing-method}
\alias{harvest,biodyn-method}
\alias{hrate}
\alias{hrate,aspic-method}
\alias{hrate,biodyn-method}
\title{harvest}
\usage{
\S4method{harvest}{biodyn,missing}(object)

hrate(object)
}
\arguments{
\item{object}{either \emph{biodyn} or \emph{aspic} classes}

\item{object}{either \emph{biodyn} or \emph{aspic} classes}
}
\description{
harvest rate

harvest rate
}

