% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrate.R
\name{hrate,biodyn-method}
\alias{hrate,biodyn-method}
\alias{hrate}
\alias{hrate,aspic-method}
\title{hrate}
\usage{
\S4method{hrate}{biodyn}(object)
}
\arguments{
\item{object}{either \emph{biodyn} or \emph{aspic} classes}
}
\description{
harvest rate
}
