% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/aspic-fit.R
\docType{methods}
\name{jk}
\alias{jk}
\alias{jk,aspic-method}
\alias{jk,biodyn,FLQuant-method}
\alias{jk,biodyn,FLQuants-method}
\title{jk}
\usage{
jk(object, index, ...)

\S4method{jk}{aspic}(object, dir = tempdir(), package = "mpb",
  exeNm = "aspic")
}
\arguments{
\item{object}{either \emph{biodyn} or \emph{aspic} class}

\item{index}{with relative abundance of type \emph{FLQuantJK} or \emph{FLQuantJKs}, if object is of type \emph{biodyn}}

\item{...}{any other parameter}

\item{object;}{an \code{aspic} object or}

\item{object;}{a character string giving an aspic "inp" file}

\item{dir;}{an optional \code{dir} where aspic text files used for fitting can be found}
}
\value{
An aspic object with fitted values and parameter estimates
}
\description{
A generic method for jack knifing a fit to catch and index of relative abundance for both \emph{biodyn} and \emph{aspic}.

Fits the aspic model to catch and catch per unit effort data removing 1 cpue observation at a time
}
\examples{
\dontrun{
    data(asp)
    asp=jk(asp)}
}
\seealso{
\code{\link{aspic}}, \code{\link{biodyn}}, \code{\link{fit}}, \code{\link{boot}}

\code{\link{biodyn},\link{boot},\link{fit}}
}

