% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{msy}
\alias{bmsy,FLPar-method}
\alias{bmsy,aspic-method}
\alias{bmsy,biodyn-method}
\alias{fmsy,FLPar-method}
\alias{fmsy,aspic-method}
\alias{fmsy,biodyn-method}
\alias{msy}
\alias{msy,FLPar-method}
\alias{msy,aspic-method}
\alias{msy,biodyn-method}
\title{msy reference points}
\usage{
msy(object, ...)
}
\arguments{
\item{object}{an object of class \code{biodyn} or \code{FLPar} with Pella-Tomlinson production
function parameters}

\item{...}{any other parameters}
}
\value{
an \code{FLPar} object with benchmark
}
\description{
Benchmarks for yield, biomass and harvest rate
}
\details{
Calculates maximum sustainable yield (MSY) reference points given the model parameters,
}
\examples{
\dontrun{ msy('logistic',FLPar(msy=100,k=500))}
  
}
\seealso{
\code{\link{refpts}}
}

