% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-msy.R
\name{msy}
\alias{msy}
\alias{msy,biodyn,missing-method}
\alias{msy,biodyn,numeric-method}
\alias{msy,character,FLPar-method}
\alias{msy,factor,FLPar-method}
\title{msy}
\usage{
msy(object, params, ...)
}
\arguments{
\item{object}{an object of class \code{biodyn} or a string or factor that species the model}

\item{params}{an \code{FLPar} object with model parameters}

\item{...}{any other parameters}
}
\value{
an \code{FLPar} object with value(s) of MSY
}
\description{
Calculates MSY given the model parameters, can be done for a biodyn class, or by specifying the model and parameters
}
\examples{
\dontrun{ msy('logistic',FLPar(msy=100,k=500))}
  
}
\seealso{
\code{\link{bmsy}}, \code{\link{fmsy}}
}

