% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/biodyn-msy.R
\docType{methods}
\name{fmsy}
\alias{bmsy}
\alias{bmsy,FLPar-method}
\alias{bmsy,aspic-method}
\alias{bmsy,biodyn-method}
\alias{bmsy-method}
\alias{fmsy}
\alias{fmsy,FLPar-method}
\alias{fmsy,aspic-method}
\alias{fmsy,biodyn-method}
\alias{fmsy-method}
\alias{msy}
\alias{msy,FLPar-method}
\alias{msy,aspic,missing-method}
\alias{msy,aspic-method}
\alias{msy,biodyn,missing-method}
\alias{msy,biodyn-method}
\alias{msy-method}
\alias{refptSD,biodyn,missing-method}
\alias{refptSD,character,FLPar-method}
\alias{refptSD,factor,FLPar-method}
\alias{refpts,FLPar,missing-method}
\alias{refpts,aspic,missing-method}
\alias{refpts,biodyn,missing-method}
\alias{refpts,character,FLPar-method}
\alias{refpts,factor,FLPar-method}
\title{msy reference points}
\usage{
fmsy(object, ...)

\S4method{msy}{biodyn,missing}(object, params = params(object), ...)
}
\arguments{
\item{object}{an object of class \code{biodyn} or \code{FLPar} with Pella-Tomlinson production
function parameters}

\item{...}{any other parameters}

\item{\code{object}}{a \code{biodyn} object}
}
\value{
an \code{FLPar} object with benchmark

an \code{FLPar} object with an estimate for $msy$
}
\description{
Calculates maximum sustainable yield (MSY) reference points given the model parameters, for yield, 
biomass and harvest rate

Calculates msy based reference points
}
\examples{
\dontrun{ msy('logistic',FLPar(msy=100,k=500))}
  
\dontrun{
 msy(bd)
 }
}

