% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-nll.R
\docType{methods}
\name{nll}
\alias{nll}
\title{nll}
\usage{
nll(object, index, ...)
}
\arguments{
\item{object}{an \code{FLQuant} with a time series of catch, iters must be equal to 1}

\item{params}{an \code{FLPar} object with parameters for the production function, r, k, p and b0.}
}
\value{
a \code{FLPar} a subset of params with parameter values that can explain the catch
}
\description{
Checks that the parameters can be taken given the catch time series
}
\examples{
\dontrun{
params=nllFn(catch,params)
}
}

