% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{diags}
\alias{diags}
\alias{setParams<-,biodyn,data.frame-method}
\alias{setParams<-,biodyn,FLPar-method}
\alias{setParams<-,biodyn,FLQuant-method}
\alias{setParams<-,biodyn,FLQuants-method}
\alias{setParams<-,aspic,data.frame-method}
\title{params}
\usage{
diags(object)
}
\arguments{
\item{object}{either \emph{biodyn} or \emph{aspic} class}

\item{value}{}

\item{...}{any other parameter}
}
\description{
Slot for estimates
}
