% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{index}
\alias{index}
\alias{setParams<-}
\alias{setParams<-,aspic,data.frame-method}
\alias{setParams<-,biodyn,FLBRP-method}
\alias{setParams<-,biodyn,FLPar-method}
\alias{setParams<-,biodyn,FLQuant-method}
\alias{setParams<-,biodyn,FLQuants-method}
\alias{setParams<-,biodyn,data.frame-method}
\title{params}
\usage{
index(object, ...)
}
\arguments{
\item{object}{either \emph{biodyn} or \emph{aspic} class}

\item{...}{any other parameter}

\item{value}{}
}
\description{
Slot for estimates
}

