% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndex.R
\name{plotIndex,FLQuants-method}
\alias{plotIndex,FLQuants-method}
\alias{plotIndex,biodyn,missing-method}
\alias{plotIndex,aspic,missing-method}
\alias{plotIndex,FLQuants,missing-method}
\title{plotIndex}
\usage{
\S4method{plotIndex}{FLQuants}(data, ...)
}
\arguments{
\item{data}{an object of class \code{FLQuants} \code{biodyn} \code{aspic}}
}
\value{
an \code{ggplot2} object
}
\description{

}
\examples{
\dontrun{

} 
}
