% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-plot.R
\docType{methods}
\name{plotProduction,biodyn,missing-method}
\alias{plotProduction,biodyn,FLQuant-method}
\alias{plotProduction,biodyn,missing-method}
\alias{plotProduction,biodyns,missing-method}
\title{plotProduction}
\usage{
\S4method{plotProduction}{biodyn,missing}(data, biomass = FLQuant(seq(0,
  max(params(data)["k"]), length.out = 101)), ...)
}
\arguments{
\item{biomass}{optional argument, an FLQuant with biomass at beginning of year}

\item{...}{other arguments}

\item{object}{an object of class \code{biodyn}}
}
\value{
an \code{ggplot2} object
}
\description{
Creates a \code{ggplot2} object that plots equilibrium values of biomass, harvest rate and catch against each other.
The basic object can then be modified by adding ggpot2 layers.
}
\examples{
\dontrun{
 refpts('logistic',FLPar(msy=100,k=500))
}
}
\seealso{
\code{\link{plotMSE}}, \code{\link{plotEql}}
}

