% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLBRP-production.R
\name{production}
\alias{production}
\alias{production,biodyn-method}
\title{production}
\usage{
production(object, biomass, ...)
}
\arguments{
\item{object}{either \emph{biodyn} or \emph{FLBRP} class}

\item{biomass}{an \emph{FLQuant}}

\item{...}{any other parameter}
}
\description{
Estimates production for a given biomass
}

