% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLBRP-production.R
\name{production}
\alias{production}
\title{production}
\usage{
production(object, stk, ...)
}
\arguments{
\item{object}{an object of class \code{FLBRP}}

\item{stk}{an \code{FLStock}}

\item{slots}{if provides then allows for non-stationarity by using the observed 
parameters each year, rather than the averages in  \code{FLBRP}}
}
\description{
Estimates estimates surplus production based on equilibrium analysis
}
\details{
Returns a data.frame with catch, biomass, observed (obs) and
expected (hat) production by year
}
\examples{
\dontrun{
   prd=production(x)
}
}

