% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspic-profile.R
\docType{methods}
\name{profileFn}
\alias{profileFn}
\title{profile}
\usage{
profileFn(r = 0.5, k = 1000, p = 0.001, b0 = 0.75, min = 0.1,
  max = 10)
}
\arguments{
\item{fitted:}{an \code{aspic} object}

\item{which:}{\code{character} giving the parameters to do the profile for, i.e. to fix.}

\item{range;}{\code{numeric} relative values by which to vary parameter values, default seq(0.5,1.5,length.out=21).}

\item{fn:}{\code{function} that gives values to be profiled.}

\item{run:}{\code{logical} if \code{TRUE} then returns profile, otherwise it just sets the control object-}
}
\value{
a \code{data frame} with results turned by \code{fn} by values in \code{which}.
}
\description{
Performs a profile using residual sum of squares, fixes some parameters for a range of values 
and then estimate the others
}
\seealso{
\code{\link{biodyn},\link{fit}}
}

