% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biodyn-control.R
\name{setControl<-}
\alias{setControl<-}
\alias{setControl<-,biodyn,FLPar-method}
\alias{setControl<-,biodyn,FLQuant-method}
\alias{setControl<-,biodyn,FLQuants-method}
\title{setControl<-}
\usage{
setControl(object, ...) <- value
}
\arguments{
\item{object}{an object of class \code{biodyn}}

\item{...}{any other parameter}

\item{value}{\code{params} object}
}
\description{
Sets the control slot in a biodyn object given the parameters in the \code{params}
slot. The starting values \code{val} are set to those in \code{params} and the \code{min} and
\code{max} bounds to .1 and 10 times these.
}
\examples{
\dontrun{
bd=sim()
setControl(bd)=params(bd)
}
}
\seealso{
\code{\link{controlFn}}
}

