% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-sim.R
\name{sim,missing,missing-method}
\alias{sim,missing,missing-method}
\alias{sim-method}
\alias{sim,FLStock,ANY-method}
\title{sim}
\usage{
\S4method{sim}{missing,missing}(
  params = FLPar(r = 0.5, k = 1000, p = 1, b0 = 1),
  harvest = FLQuant(c(seq(0, 1.5, length.out = 30), rev(seq(0.5, 1.5, length.out =
    15))[-1], rep(0.5, 5)) * (params["r"] * (1/(1 + params["p"])))),
  bounds = c(0.1, 10),
  p = NULL,
  b0 = NULL,
  ...
)
}
\arguments{
\item{params}{surplus production parameters}

\item{harvest}{\code{FLQuant} with harvest rate}

\item{bounds}{on \code{control}}

\item{...}{other arguments}

\item{model}{character corresponding to model}
}
\value{
biodyn object with simulated time series
}
\description{
Creates a biodyn object with known properties
}
\examples{
\dontrun{
 bd=sim() 
 }
}
