% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodyn-methods.R
\name{computePrd}
\alias{computePrd}
\alias{computePrd,biodyn,FLQuant-method}
\alias{computePrd,biodyn,missing-method}
\alias{computePrd,biodyn,numeric-method}
\title{computePrd}
\usage{
computePrd(object, biomass, ...)
}
\arguments{
\item{object}{an object of class \code{biodyn}}

\item{biomass}{stock biomaas, may be a \code{numerix},  \code{FLQuant} or missing. In the latte case the stock slot will be used.}

\item{...}{other arguments}
}
\value{
an \code{FLPar} object
}
\description{
Calculates the surplus production for a biomass dynamic model given a level of stock biomass
}
\examples{
\dontrun{ computePrd(bd,seq(0,params(bd)['k'])) }
 
}
\seealso{
\code{\link{plotPrd}}
}

