% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mng-methods.R
\name{fapexAge}
\alias{fapexAge}
\alias{timeSeries}
\alias{timeSeries-method}
\alias{timeSeries,FLStock,FLBRP-method}
\alias{timeSeries,FLStocks,FLBRP-method}
\alias{timeSeries,FLStocks,FLBRPs-method}
\alias{timeSeries,aspic,missing-method}
\alias{timeSeries,biodyn,missing-method}
\alias{aspic,missing}
\alias{biodyn,missing}
\title{mng}
\usage{
fapexAge(object)
}
\arguments{
\item{object}{either  \emph{FLStock} \emph{biodyn} or \emph{aspic} classes}
}
\description{
calculates time series of quantities useful for management
}
